Abstracted out the Character and Int extensions since they are not exclusive to the Caesar cipher and are used for other shift ciphers
